+++ packages/litqa/src/aviary/envs/litqa/__init__.py
@@ -1,10 +1,12 @@
 from .env import (
+    DEFAULT_REWARD_MAPPING,
+    make_discounted_returns,
+)
+from .task import (
     DEFAULT_AVIARY_PAPER_HF_HUB_NAME,
     DEFAULT_LABBENCH_HF_HUB_NAME,
-    DEFAULT_REWARD_MAPPING,
     LitQAv2TaskDataset,
     LitQAv2TaskSplit,
-    make_discounted_returns,
     read_litqa_v2_from_hub,
 )
