+++ packages/litqa/src/aviary/envs/litqa/env.py
@@ -1,8 +1,7 @@
 import sys
-from collections.abc import Awaitable, Callable, Mapping, Sequence
+from collections.abc import Awaitable, Callable, Mapping
 
 from aviary.utils import (
-    DEFAULT_EVAL_MODEL_NAME,
     MultipleChoiceEvaluation,
     MultipleChoiceQuestion,
 )
@@ -13,39 +12,25 @@ else:
     from typing_extensions import TypeVar  # For TypeVar.default backport
 
 import logging
-import re
-from abc import ABC
 from copy import deepcopy
-from enum import StrEnum
-from typing import Any, Generic, Self, assert_never, cast
+from typing import Any, Generic, Self, cast
 from uuid import UUID
 
-import pandas as pd
-from ldp.data_structures import Trajectory
-from llmclient import EmbeddingModel, LiteLLMModel, LLMModel
+from llmclient import EmbeddingModel, LiteLLMModel
 from paperqa.agents.env import POPULATE_FROM_SETTINGS, PaperQAEnvironment
 from paperqa.agents.search import SearchIndex, maybe_get_manifest
-from paperqa.agents.tools import Complete
 from paperqa.docs import Docs
 from paperqa.settings import Settings
-from paperqa.types import DocDetails
 
 from aviary.core import (
-    TASK_DATASET_REGISTRY,
     Messages,
-    TaskDataset,
     ToolRequestMessage,
-    ToolResponseMessage,
 )
 
 logger = logging.getLogger(__name__)
 TEvaluation = TypeVar("TEvaluation", default=MultipleChoiceEvaluation)
 
-
 DEFAULT_REWARD_MAPPING = {"correct": 1.0, "unsure": 0.1, "incorrect": -1.0}
-DEFAULT_LABBENCH_HF_HUB_NAME = "futurehouse/lab-bench"
-# Test split from Aviary paper's section 4.3: https://doi.org/10.48550/arXiv.2412.21154
-DEFAULT_AVIARY_PAPER_HF_HUB_NAME = "futurehouse/aviary-paper-data"
 
 
 def discounted_returns(
@@ -252,231 +237,3 @@ class GradablePaperQAEnvironment(PaperQAEnvironment, Generic[TEvaluation]):
         # tool functions within the tools
         copy_self.tools = copy_self.make_tools()
         return copy_self
-
-
-def read_litqa_v2_from_hub(
-    train_eval_dataset: str = DEFAULT_LABBENCH_HF_HUB_NAME,
-    test_dataset: str = DEFAULT_AVIARY_PAPER_HF_HUB_NAME,
-    randomize: bool = True,
-    seed: int | None = None,
-    train_eval_split: float = 0.8,
-) -> tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
-    """
-    Read LitQA v2 JSONL into train, eval, and test DataFrames.
-
-    Args:
-        train_eval_dataset: Hugging Face Hub dataset's name corresponding with train
-            and eval splits.
-        test_dataset: Hugging Face Hub dataset's name corresponding with a test split.
-        randomize: Opt-out flag to shuffle the dataset after loading in by question.
-        seed: Random seed to use for the shuffling.
-        train_eval_split: Train/eval split fraction, default is 80% train 20% eval.
-
-    Raises:
-        DatasetNotFoundError: If any of the datasets are not found, or the
-            user is unauthenticated.
-    """
-    try:
-        from datasets import load_dataset
-    except ImportError as exc:
-        raise ImportError(
-            "Reading in LitQA2 requires the 'datasets' extra for 'datasets'. Please:"
-            " `pip install paper-qa[datasets]`."
-        ) from exc
-
-    train_eval = load_dataset(train_eval_dataset, "LitQA2")["train"].to_pandas()
-    test = load_dataset(test_dataset, "LitQA2")["test"].to_pandas()
-    # Convert to list so it's not unexpectedly a numpy array
-    train_eval["distractors"] = train_eval["distractors"].apply(list)
-    test["distractors"] = test["distractors"].apply(list)
-    # Let downstream usage in the TaskDataset's environment factories check for the
-    # presence of other DataFrame columns
-    if randomize:
-        train_eval = train_eval.sample(frac=1, random_state=seed)
-        test = test.sample(frac=1, random_state=seed)
-    num_train = int(len(train_eval) * train_eval_split)
-    return train_eval[:num_train], train_eval[num_train:], test
-
-
-class LitQATaskDataset(
-    TaskDataset[GradablePaperQAEnvironment], ComputeTrajectoryMetricsMixin, ABC
-):
-    """
-    Abstract base class for a task dataset of LitQA v1 or v2 questions.
-
-    This is an ABC because it's non-specific to a LitQA version.
-    Examples include LitQA v1, v2, or a test stub version of LitQA.
-    """
-
-    def __init__(
-        self,
-        settings: Settings | dict | None = None,
-        base_docs: Docs | dict | None = None,
-        rewards: Mapping[str, float] = DEFAULT_REWARD_MAPPING,
-        question_kwargs: Mapping[str, Any] | None = None,
-        eval_model: LLMModel | str = DEFAULT_EVAL_MODEL_NAME,
-        **env_kwargs,
-    ):
-        if settings is None:
-            settings = Settings()
-        if isinstance(settings, dict):
-            settings = Settings(**settings)
-        self._settings = settings
-        if base_docs is None:
-            base_docs = Docs()
-        if isinstance(base_docs, dict):
-            base_docs = Docs(**base_docs)
-        self._base_docs = base_docs
-        self._rewards = rewards
-        self._question_kwargs = question_kwargs
-        self._eval_model = eval_model
-        self._env_kwargs = env_kwargs
-
-    def _make_gradable_environment(
-        self,
-        ideal_answer: str,
-        distractors: str | list[str],
-        question: str,
-        sources: str | list[str] | None = None,
-    ) -> GradablePaperQAEnvironment:
-        mc_question = MultipleChoiceQuestion(
-            question=question,
-            options=(
-                distractors
-                if isinstance(distractors, list)
-                else MultipleChoiceQuestion.split_options(distractors)
-            ),
-            ideal_answer=ideal_answer,
-            **(self._question_kwargs or {}),
-        )
-        return GradablePaperQAEnvironment(
-            query=mc_question,
-            settings=self._settings,
-            docs=self._base_docs.model_copy(),
-            sources=sources,
-            rewards=self._rewards,
-            **self._env_kwargs,
-        )
-
-    def compute_trajectory_metrics(
-        self, trajectories: "Sequence[Trajectory]"
-    ) -> dict[str, list[float]]:
-        total_paper_count: list[float] = []
-        relevant_paper_count: list[float] = []
-        evidence_count: list[float] = []
-        for t in trajectories:
-            split_certainties = [
-                split_certainty
-                for split_certainty in (
-                    re.split(
-                        pattern=Complete.CERTAINTY_SPLIT_REGEX_PATTERN,
-                        string=obs.content,
-                        maxsplit=1,
-                    )
-                    for obs in t.steps[-1].next_observation
-                    if (
-                        isinstance(obs, ToolResponseMessage)
-                        and obs.name == Complete.TOOL_FN_NAME
-                    )
-                )
-                # Filter for places where the regex split succeeded
-                if len(split_certainty) >= 4  # noqa: PLR2004
-            ]
-            for i, metric_list in enumerate(
-                (total_paper_count, relevant_paper_count, evidence_count),
-                start=1,  # Regex extraction of status starts after has_successful_answer
-            ):
-                # NOTE: we use mean to not break if there's 2+ complete calls (which
-                # we're prompted not to do). If it happens, they should all have the
-                # same status, so the mean value should equal the individual values
-                metric_list.append(
-                    sum(int(sa[i]) for sa in split_certainties) / len(split_certainties)
-                    if split_certainties  # Avoid div0 (when complete wasn't called)
-                    else 0
-                )
-        return super().compute_trajectory_metrics(trajectories) | {
-            "total_paper_count": total_paper_count,
-            "relevant_paper_count": relevant_paper_count,
-            "evidence_count": evidence_count,
-            "correct": [
-                int(t.steps[-1].reward == self._rewards["correct"])
-                for t in trajectories
-            ],
-            "correct_unsure": [
-                int(
-                    t.steps[-1].reward
-                    in {self._rewards["correct"], self._rewards["unsure"]}
-                )
-                for t in trajectories
-            ],
-        }
-
-
-class LitQAv2TaskSplit(StrEnum):
-    TRAIN = "train"
-    EVAL = "eval"
-    TEST = "test"
-
-    def get_index(self) -> int:
-        """
-        Get the index of the train (0), eval (1), or test (2) split.
-
-        NOTE: the value matches the index in read_litqa_v2_from_hub's returned splits.
-        """
-        if self == self.TRAIN:
-            return 0
-        if self == self.EVAL:
-            return 1
-        if self == self.TEST:
-            return 2
-        assert_never(self)  # type: ignore[arg-type]
-
-
-class LitQAv2TaskDataset(LitQATaskDataset):
-    """Task dataset of LitQA v2 questions."""
-
-    def __init__(
-        self,
-        *args,
-        train_eval_dataset: str = DEFAULT_LABBENCH_HF_HUB_NAME,
-        test_dataset: str = DEFAULT_AVIARY_PAPER_HF_HUB_NAME,
-        read_data_kwargs: Mapping[str, Any] | None = None,
-        split: str | LitQAv2TaskSplit = LitQAv2TaskSplit.EVAL,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-        split_dfs = read_litqa_v2_from_hub(
-            train_eval_dataset, test_dataset, **(read_data_kwargs or {})
-        )
-        self.data = split_dfs[LitQAv2TaskSplit(split).get_index()]
-
-    def get_new_env_by_idx(self, idx: int) -> GradablePaperQAEnvironment:
-        sources = []
-        for s in self.data.iloc[idx].sources:
-            try:
-                (doi,) = (
-                    s.split(substr, maxsplit=1)[1]
-                    for substr in DocDetails.DOI_URL_FORMATS
-                    if substr in s
-                )
-            except ValueError as exc:
-                raise NotImplementedError(
-                    f"Didn't handle DOI extraction from source {s!r}."
-                ) from exc
-            sources.append(doi)
-        return self._make_gradable_environment(
-            ideal_answer=self.data.iloc[idx].ideal,
-            distractors=self.data.iloc[idx].distractors,
-            question=self.data.iloc[idx].question,
-            sources=sources,
-        )
-
-    def __len__(self) -> int:
-        return len(self.data)
-
-
-TASK_DATASET_NAME = "litqa-v2"
-TASK_DATASET_REGISTRY[TASK_DATASET_NAME] = (
-    LitQAv2TaskDataset.__module__,
-    LitQAv2TaskDataset.__name__,
-)
